requirements.txt
appdirs==1.4.3
asgiref==3.3.4
asn1crypto==0.24.0
awscli==1.14.44
Babel==2.6.0
beautifulsoup4==4.9.3
botocore==1.16.19
Brlapi==0.6.6
catfish==1.4.4
certifi==2020.12.5
chardet==3.0.4
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cssmin==0.2.0
cssselect==1.0.3
cssutils==1.0.2
cupshelpers==1.0
decorator==4.3.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.14
ebaysdk==2.2.0
expeyes==1.0.0
fail2ban==0.10.2
feedparser==6.0.2
filelock==3.0.12
flake8==3.5.0
gevent==20.12.1
greenlet==1.0.0
gunicorn==20.1.0
html2text==2020.1.16
html5lib==0.999999999
httplib2==0.9.2
idna==2.10
importlib-metadata==1.5.0
importlib-resources==1.3.1
Jinja2==2.11.2
jmespath==0.9.3
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
libsass==0.20.1
louis==3.5.0
lxml==4.6.2
macaroonbakery==1.1.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.3
netifaces==0.10.4
num2words==0.5.10
numpy==1.13.3
oauth==1.0.1
ofxparse==0.20
olefile==0.45.1
onboard==1.4.1
openshot-qt==2.4.1
passlib==1.7.4
pexpect==4.2.1
Pillow==8.1.0
polib==1.1.0
protobuf==3.0.0
psutil==5.8.0
psycogreen==1.0.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
pyenchant==2.0.0
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==2.1.0
pyldap==3.0.0.post1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.0
pyserial==3.5
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-espeak==0.5
python-ldap==3.3.1
python-openid==2.2.5
python-sane==2.8.3
python-stdnum==1.15
pytz==2021.1
pyusb==1.1.0
pyxdg==0.25
PyYAML==5.3.1
pyzmq==16.0.2
qrcode==6.1
regex==2017.12.12
reportlab==3.5.59
requests==2.25.1
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.3.3
scipy==0.19.1
scour==0.36
SecretStorage==2.3.1
sgmllib3k==1.0.0
simplejson==3.13.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
ssh-import-id==5.7
suds-jurko==0.6
system-service==0.3
systemd-python==234
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.3
vatnumber==1.2
virtualenv==20.0.9
vobject==0.9.6.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.11.15
whitenoise==5.2.0
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
